apps/corde/engine/requirements.txt
# CORDE Engine Dependencies
# Install: pip install -r requirements.txt

# Core ML
torch>=2.1.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# IP-Adapter (when available)
# ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git

# Audio processing (for VideoPoesia / music-sync)
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
